Basic setting loading works. Display mode isn't being set correctly
The settings can be loaded from the settings file, and resets the
file if it doesn't exist.

The display resolution can go to any custom resolution.
Instead it should search for the correct resolution from
a list of supported ones. If it doesn't find one, it should go to
a default one.
Signed-off-by: Kolijn Wolfaardt <kolijn.wolfaardt@gmail.com>